Ganz großes refactoring. Auslagern von den Komponenten im Frontend, richtiges zusammensetzten der einzelnen Komponenten. LIME funktioniert leider noch nicht optimal, und SHAP und LIME funktionieren für LR und GAM noch nicht
